Fix warnings in hwiu test app
So it's easier to see warnings about actual
problems with new changes.

+++ libs/hwui/tests/main.cpp
@@ -111,7 +111,7 @@ int main(int argc, char* argv[]) {
 
     for (int i = 0; i < 150; i++) {
         ATRACE_NAME("UI-Draw Frame");
-        for (int ci = 0; ci < cards.size(); ci++) {
+        for (size_t ci = 0; ci < cards.size(); ci++) {
             cards[ci]->mutateStagingProperties().setTranslationX(i);
             cards[ci]->mutateStagingProperties().setTranslationY(i);
             cards[ci]->setPropertyFieldsDirty(RenderNode::X | RenderNode::Y);
